test: share Google ID token fixture
